requirements.txt
fpdf==1.7.2
numpy==1.19.2
pandas==1.1.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
XlsxWriter==1.3.6
xlrd==1.2.0
black==20.8b1
